requirements.txt
# requirements.txt
# Install with: pip install -r requirements.txt

fastapi
uvicorn[standard]
python-dotenv
openai
# Pipecat and dependencies
pipecat-ai==0.0.100
pipecat-ai[silero]
daily-python==0.23.0


# AI Services

#anthropic==0.18.1  # If using Claude instead of OpenAI

# Speech Services
# deepgram-sdk==3.2.0
# cartesia==1.0.0
# elevenlabs==0.2.27  # Alternative TTS option

# Voice Activity Detection
# torch
# torchaudio

